# swagger.yml
swagger: "2.0"
info:
  description: "URL shortener app"
  version: "1.0.0"
  title: "Shortener"
  contact:
    email: "tianyizhang1987@gmail.com"
  license:
    name: "ISC"
host: "tbd.com"
basePath: "/api"
schemes:
- "https"
- "http"
paths:
  /shorten:
    post:
      summary: "Submit a long url to shorten it"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "url_to_shorten"
        description: "Long url that needs to be shortened"
        required: true
        schema:
          type: "string"
      responses:
        200:
          description: "Successful operation, url has been shortened"
          schema:
            type: "object"
            properties: 
              id: 
                type: "integer"
                format: "int64"
              shortened_url:
                type: "string"
        400: 
          description: "Invalid long url input"
        405:
          description: "Invalid method"
  /{code}:
    get: 
      summary: "End point for redirecting to website of original long url"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "unique_code"
        in: "path"  
        description: "Code of the short url"
        required: true
        type: "string"
      responses: 
        302:
          description: "Successful redirect"
        400: 
          description: "Invalid unique_code for shortened url"
  /urls/{unique_code}:
    get: 
      summary: "End point for retrieving shortened url, original long url, and row id by unique_code"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "unique_code"
        in: "path"  
        description: "Code of the short url"
        required: true
        type: "string"
      responses: 
        200:
          description: "Successful retrieval of url object"
          schema:
            type: "object"
            properties:
              id:
                type: "integer"
                format: "int64"
              shortened_url:
                type: "string"
              original_url:
                type: "string"
        400: 
          description: "Invalid unique_code for shortened url"
        405:
          description: "Invalid method"
definitions:
  url_item:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      original_url:
        type: "string"
      unique_code:
        type: "string"